% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.clust_fcts.R
\name{bootNSD}
\alias{bootNSD}
\title{Iterative latent-state model of NSD}
\usage{
bootNSD(x, y, day, interval = 30, by = 5, n.iter = 20000, ...)
}
\arguments{
\item{x, }{y, time,  information on locations and time (continuous)}

\item{interval}{Amount of variation in time to be tested}

\item{by}{Time step between each iteration (number of iterations is a function of the interval and time step)}

\item{n.iter}{number of total iterations per chain (including burn in; default: 5000)}
}
\value{
A boot.clust object
}
\description{
Bootstrap latent-state model by changing starting locations
}
\examples{
data(Christian)
#May results in low convergence given low number of iterations
boot1<-bootNSD(Christian$x[1:300], Christian$y[1:300], Christian$Time[1:300], interval=20, by=5, n.iter=10000) 
summary(boot1)
plot(boot1)
}
\keyword{boot.clust}
\keyword{bootNSD}
\keyword{clustNSD}

