% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust.classify_fcts.R
\name{breaks}
\alias{breaks}
\title{Calculating transitions between clusters}
\usage{
breaks(object)
}
\arguments{
\item{object}{an object of class "mov.clust" or "rjags"}
}
\description{
Calculating pattern of transitions among cluster
}
\examples{
data(Zelfa_rjags)
breaks(Zelfa_rjags)		
}
\keyword{classify}
\keyword{clust.classify}

