% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{Christian}
\alias{Christian}
\title{Christian - daily spatial locations}
\format{A data frame with 1536 rows and 3 variables (x, y, time)}
\usage{
Christian
}
\description{
A dataset containing the x, y locations (UTM), and time of Christian, a male giant tortoises inhabiting Alcedo volcano on Isabela Island, Galapagos.
}
\keyword{datasets}

