% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{Christian_rjags}
\alias{Christian_rjags}
\title{Latent-state model applied to Christian}
\format{A "rjags" object}
\usage{
Christian_rjags
}
\description{
The output of clustNSD function applied to Christian.
}
\keyword{datasets}

