% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust.classify_fcts.R
\name{classify}
\alias{classify}
\title{Classify movement strategies}
\usage{
classify(object, grph = F, ...)
}
\arguments{
\item{object}{an object of class "mov.clust" or "rjags"}

\item{grph}{Produce a graph of classification (Default = F)}
}
\value{
A clust.classify object
}
\description{
Extract information to classify movement strategies
}
\examples{
data(Christian_rjags)
summary(classify(Christian_rjags))	
data(Zelfa_rjags)
summary(classify(Zelfa_rjags))
}
\keyword{classify}
\keyword{clust.classify}
\keyword{mov.clust}

