% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_fcts.R
\name{colVars}
\alias{colVars}
\title{Posterior variances}
\usage{
colVars(a)
}
\arguments{
\item{a}{A matrix}
}
\description{
This function calculates posterior variances from simulation. Based on script initially written by Andrew Gelman
}
\examples{
mat<-matrix(rnorm(20), nrow=10, ncol=2)
}
\keyword{waic}

