% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_fcts.R
\name{Dist_fct}
\alias{Dist_fct}
\title{Calculate euclidean distance}
\usage{
Dist_fct(x, y)
}
\arguments{
\item{x, y}{Spatial locations}
}
\description{
This function calculates euclidean distance between spatial locations
}
\examples{
data(Christian)
hist(Dist_fct(Christian$x, Christian$y))
}
\keyword{distance}

