% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{package}
\name{lsmnsd}
\alias{lsmnsd}
\alias{lsmnsd-package}
\title{lsmnsd: A package for classifying animal movement strategies based on a latent-space model and net squared displacement.}
\description{
The lsmnsd package provides four categories of important functions:
clustNSD, simple.clust, classify, bootNSD.
}

