% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_fcts.R
\name{Mode}
\alias{Mode}
\title{Mode}
\usage{
Mode(x)
}
\arguments{
\item{x}{A vector}
}
\description{
This function finds the mode of a vector
}
\examples{
v<-c(1,2,2,3,4,4,4,4,5,6)
Mode(v)
}
\keyword{mode}

