% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_fcts.R
\name{NSD_fct}
\alias{NSD_fct}
\title{Calculate NSD}
\usage{
NSD_fct(x, y)
}
\arguments{
\item{x, y}{Spatial locations}
}
\description{
This function calculates NSD from a time-series of spatial locations
}
\examples{
data(Christian)
ts.plot(NSD_fct(Christian$x, Christian$y))
}
\keyword{nsd}

