% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.clust_fcts.R
\name{plot.boot.clust}
\alias{plot.boot.clust}
\title{Plotting boot.clust object}
\usage{
\method{plot}{boot.clust}(object)
}
\arguments{
\item{object}{an object of class "boot.clust"}
}
\description{
plot method for class "boot.clust"
}
\examples{
#plot(boot1)
}
\keyword{boot.clust}
\keyword{bootNSD}

