% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mov.clust_fcts.R
\name{plot.mov.clust}
\alias{plot.mov.clust}
\title{Plot mov.clust object}
\usage{
\method{plot}{mov.clust}(object)
}
\arguments{
\item{object}{an object of class "mov.clust"}
}
\description{
plotting method for class "mov.clust"
}
\examples{
data(Zelfa_rjags)
plot(simple.clust(Zelfa_rjags))
}
\keyword{mov.clust}
\keyword{simple.clust}

