% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mov.clust_fcts.R
\name{print.mov.clust}
\alias{print.mov.clust}
\title{Print mov.clust object}
\usage{
\method{print}{mov.clust}(object)
}
\arguments{
\item{object}{an object of class "mov.clust"}
}
\description{
printing method for class "mov.clust"
}
\examples{
data(Zelfa_rjags)
simple.clust(Zelfa_rjags)
}
\keyword{mov.clust}
\keyword{simple.clust}

