% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_fcts.R
\name{range01}
\alias{range01}
\title{Range standardisation (0,1)}
\usage{
range01(x)
}
\arguments{
\item{x}{A vector}
}
\description{
This function standardises a vector between 0 and 1
}
\examples{
v<-c(1,2,2,3,4,4,5,6)
range01(v)
}
\keyword{range}
\keyword{standardisation}

