% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjags_fcts.R
\name{rerun}
\alias{rerun}
\title{Rerun clustNSD}
\usage{
rerun(rjags, n.update = 3, inc.fact = 1)
}
\arguments{
\item{rjags}{Output from clustNSD}

\item{n.update}{the max number of updates, default =3}

\item{inc.fact}{Factor by which the number of iterations will be increased, default=1 (no increase)}
}
\description{
This function uses the autojags function to rerun the function clustNSD to facilitate convergence
}
\examples{
# DO NOT RUN - TAKES A LONG TIME
#data(Christian_rjags)
#rerun(Christian_rjags, n.update=2, inc.fact=1)
}
\keyword{rerun}

