% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mov.clust_fcts.R
\name{simple.clust}
\alias{simple.clust}
\title{Convert to mov.clust class}
\usage{
simple.clust(out)
}
\arguments{
\item{out}{Output of clustNSD function}
}
\value{
A mov.clust object
}
\description{
This function simplifies a rjags output from clustNSD to a mov.clust object
}
\examples{
data(Christian_rjags)
summary(simple.clust(Christian_rjags))
plot(simple.clust(Christian_rjags))
data(Zelfa_rjags)
summary(simple.clust(Zelfa_rjags))
plot(simple.clust(Zelfa_rjags))
}
\keyword{mov.clust}
\keyword{simple.clust}

