% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.clust_fcts.R
\name{summary.boot.clust}
\alias{summary.boot.clust}
\title{Summarizing boot.clust object}
\usage{
\method{summary}{boot.clust}(object)
}
\arguments{
\item{object}{an object of class "boot.clust"}
}
\description{
summary method for class "boot.clust"
}
\examples{
summary(boot1)
}
\keyword{boot.clust}
\keyword{bootNSD}

