% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust.classify_fcts.R
\name{summary.clust.classify}
\alias{summary.clust.classify}
\title{Summarizing clust.classify object}
\usage{
\method{summary}{clust.classify}(object)
}
\arguments{
\item{object}{an object of class "clust.classify"}
}
\description{
summary method for class "clust.classify"
}
\examples{
data(Christian_rjags)
summary(classify(Christian_rjags))	
}
\keyword{classify}
\keyword{clust.classify}

