% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mov.clust_fcts.R
\name{summary.mov.clust}
\alias{summary.mov.clust}
\title{Summarizing mov.clust object}
\usage{
\method{summary}{mov.clust}(object)
}
\arguments{
\item{object}{an object of class "mov.clust"}
}
\description{
summary method for class "mov.clust"
}
\examples{
data(Christian_rjags)
summary(simple.clust(Christian_rjags))
}
\keyword{mov.clust}
\keyword{simple.clust}

