% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust.classify_fcts.R
\name{switch.matrix}
\alias{switch.matrix}
\title{Switching probabilties matrix}
\usage{
switch.matrix(object)
}
\arguments{
\item{object}{an object of class "mov.clust" or "rjags"}
}
\description{
Extrac matrix of switching probabilities
}
\examples{
data(Zelfa_rjags)
switch.matrix(Zelfa_rjags)	
}
\keyword{classify}
\keyword{clust.classify}

