% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust.classify_fcts.R
\name{time.spent}
\alias{time.spent}
\title{Calculating time-spent in cluster}
\usage{
\method{time}{spent}(object)
}
\arguments{
\item{object}{an object of class "mov.clust" or "rjags"}
}
\description{
Calculating pattern of time spent in each cluster
}
\examples{
data(Zelfa_rjags)
time.spent(Zelfa_rjags)
}
\keyword{classify}
\keyword{clust.classify}

