% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_fcts.R
\name{waic}
\alias{waic}
\title{WAIC calculation}
\usage{
waic(jagsfit)
}
\arguments{
\item{jagsfit}{Output of a clustNSD call}
}
\description{
This function calculates WAIC from clustNSD output. Based on script initially written by Andrew Gelman
}
\examples{
data(Christian_rjags)
waic(Christian_rjags)
}
\keyword{waic}

