% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{Zelfa}
\alias{Zelfa}
\title{Zelfa - daily spatial locations}
\format{A data frame with 1537 rows and 3 variables (x, y, time)}
\usage{
Zelfa
}
\description{
A dataset containing the x, y locations (UTM), and time of Zelfa, a female giant tortoises inhabiting Espanola Island, Galapagos.
}
\keyword{datasets}

