% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{Zelfa_rjags}
\alias{Zelfa_rjags}
\title{Latent-state model applied to Zelfa}
\format{A "rjags" object}
\usage{
Zelfa_rjags
}
\description{
The output of clustNSD function applied to Zelfa.
}
\keyword{datasets}

